added new commands (update, delete), and query handler, updated the GenericRepository and registered MediatR in API

+++ TaskFlow/TaskFlow.Application/Interfaces/Repository/IGenericRepository.cs
@@ -9,8 +9,8 @@ namespace TaskFlow.Application.Interfaces.Repository
     public interface IGenericRepository<TEntity> where TEntity : class
     {
         public Task AddGeneric(TEntity user);
-        public Task<bool> DeletByIdGenericAsync(int Id);
-        public Task<TEntity> GetEntityById(int Id);
+        public Task<bool> DeletByIdGenericAsync(ushort Id);
+        public Task<TEntity> GetEntityById(ushort Id);
         public Task<IEnumerable<TEntity>> GetAllEntities();
     }
 }
